Complete change of EntityData (added generic entity attributes)

+++ src/main/java/mixac1/dangerrpg/event/LvlableItemEventHandlers.java
@@ -6,8 +6,8 @@ import cpw.mods.fml.common.gameevent.TickEvent;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import mixac1.dangerrpg.capability.LvlableItem;
-import mixac1.dangerrpg.capability.itemattr.ItemAttributes;
-import mixac1.dangerrpg.capability.playerattr.PlayerAttributes;
+import mixac1.dangerrpg.capability.ea.PlayerAttributes;
+import mixac1.dangerrpg.capability.ia.ItemAttributes;
 import mixac1.dangerrpg.event.ItemStackEvent.AddInformationEvent;
 import mixac1.dangerrpg.event.ItemStackEvent.HitEntityEvent;
 import mixac1.dangerrpg.event.ItemStackEvent.OnLeftClickEntityEvent;
@@ -33,35 +33,35 @@ public class LvlableItemEventHandlers
             }
             ((EntityLivingBase) e.entity).hurtResistantTime = 0;
         }
-        
+
         return false;
     }
-    
+
     @SubscribeEvent
     public boolean hitEntity(HitEntityEvent e)
     {
         if (e.attacker instanceof EntityPlayer) {
             EntityPlayer player = (EntityPlayer) e.attacker;
-            
+
             if (ItemAttributes.KNOCKBACK.hasIt(e.stack)) {
                 RPGCommonHelper.knockBack(e.entity, e.attacker, ItemAttributes.KNOCKBACK.get(e.stack, player));
             }
-            
+
             e.entity.hurtResistantTime = 0;
             if (ItemAttributes.STR_MUL.hasIt(e.stack)) {
                 e.entity.attackEntityFrom(DamageSource.causePlayerDamage(player),
                     PlayerAttributes.STRENGTH.getValue(player) * ItemAttributes.STR_MUL.get(e.stack));
             }
-            
+
             if (ItemAttributes.MELEE_SPEED.hasIt(e.stack)) {
                 float speed = ItemAttributes.MELEE_SPEED.get(e.stack, player);
-                PlayerAttributes.SPEED_COUNTER.setValue((int) (speed < 0 ? 0 : speed), player);
+                PlayerAttributes.SPEED_COUNTER.setValue(speed < 0 ? 0 : speed, player);
             }
         }
-        
+
         return true;
     }
-    
+
     @SubscribeEvent
     @SideOnly(Side.CLIENT)
     public void addInformation(AddInformationEvent e)
@@ -73,7 +73,7 @@ public class LvlableItemEventHandlers
                      (int) ItemAttributes.CURR_EXP.get(e.stack) + "/" + (int) ItemAttributes.MAX_EXP.get(e.stack));
         }
     }
-    
+
     @SubscribeEvent
     public void onPlayerTick(TickEvent.PlayerTickEvent e)
     {
@@ -82,7 +82,7 @@ public class LvlableItemEventHandlers
             if (e.player != null && (tmp = PlayerAttributes.SPEED_COUNTER.getValue(e.player)) > 0) {
                 PlayerAttributes.SPEED_COUNTER.setValue(tmp - 1, e.player);
             }
-            
+
             if (e.player.worldObj.getTotalWorldTime() % 20 == 0 &&
                 PlayerAttributes.CURR_MANA.getValue(e.player) < PlayerAttributes.MANA.getValue(e.player) &&
                 (tmp = PlayerAttributes.MANA_REGEN.getValue(e.player)) != 0) {
@@ -90,7 +90,7 @@ public class LvlableItemEventHandlers
             }
         }
     }
-    
+
     @SideOnly(Side.CLIENT)
     @SubscribeEvent
     public void onPlayerTickClient(TickEvent.PlayerTickEvent e)
@@ -99,12 +99,12 @@ public class LvlableItemEventHandlers
             if (e.player == null) {
                 return;
             }
-            
+
             if (e.player.swingProgressInt == 1) {
                 ItemStack stack = e.player.getCurrentEquippedItem();
-                if (stack != null && ItemAttributes.REACH.hasIt(stack)) {    
+                if (stack != null && ItemAttributes.REACH.hasIt(stack)) {
                     MovingObjectPosition object = RPGCommonHelper.getMouseOver(0, ItemAttributes.REACH.get(stack) + 4);
-                    
+
                     if (object != null && object.entityHit != null && object.entityHit != e.player && object.entityHit.hurtResistantTime == 0) {
                         FMLClientHandler.instance().getClient().playerController.attackEntity(e.player, object.entityHit);
                     }
@@ -112,7 +112,7 @@ public class LvlableItemEventHandlers
             }
         }
     }
-    
+
     @SubscribeEvent
     public void breakSpeed(BreakSpeed e)
     {
